+++ src/engine/index.js
@@ -62,7 +62,7 @@ export class FSM extends Fsm {
      * @param {*}         widget.ID                     Идентификатор
      * @param {Function} [widget.bind]                   Функция навешивающая обработчики
      * @param {Function} [widget.render]                 Функция отрисовки.
-     * @param {Function} [widget.destroy]                Функция отвязывающая обработчики
+     * @param {Function} [widget.remove]                Функция отвязывающая обработчики
      * @param {Object}   [widget.options]                Опции виджета
      * @param {Array}    [widget.options.types]          Массив типов
      * @param {Function} [widget.options.beforeRegister] До регистрации
@@ -71,8 +71,8 @@ export class FSM extends Fsm {
      * @param {Function} [widget.options.afterBind]      После навешивания обработчиков
      * @param {Function} [widget.options.beforeRender]   До отрисовки этого элемента
      * @param {Function} [widget.options.afterRender]    После отрисовки этого элемента
-     * @param {Function} [widget.options.beforeDestroy]  До отвязки обработчиков
-     * @param {Function} [widget.options.afterDestroy]   После отвязки обработчиков
+     * @param {Function} [widget.options.beforeRemove]  До отвязки обработчиков
+     * @param {Function} [widget.options.afterRemove]   После отвязки обработчиков
      * @see {@link RegistrationState#register}
      */
     register( widget ) {